Add support for wrapping checked exceptions.
If one of your methods cannot handle a checked exception by itself then
you have two options. You could rethrow the exception and force
the caller to handle the exception or you could wrap the exception with
a RuntimeException. Fishbowl's `wrapCheckedException` supports the
latter case.

    public void doSomething() {
      wrapCheckedException(() -> throw new IOException());
    }

This method may be used with statements that have a return value, too.

    public void doSomething() {
      URL url = wrapCheckedException(() -> new URL("http://junit.org/"));
      ...
    }

Only checked exceptions are wrapped. RuntimeExceptions are not wrapped.